OK-831: Refaktoroitu sijoittelun tulosten tilakoneen käyttöä komponenteissa
